<fix>[host]: extract BlockDevices from APIGetPhysicalMachineBlockDevicesReply
Resolves: ZSV-2924
