+++ header/src/main/java/org/zstack/header/host/BlockDevicesDoc_zh_cn.groovy
@@ -1,6 +1,6 @@
 package org.zstack.header.host
 
-import org.zstack.header.host.APIGetPhysicalMachineBlockDevicesReply.BlockDevices.BlockDevice
+import org.zstack.header.host.BlockDevices.BlockDevice
 
 doc {
 
@@ -8,7 +8,7 @@ doc {
 
 	ref {
 		name "unusedBlockDevices"
-		path "org.zstack.header.host.APIGetPhysicalMachineBlockDevicesReply.BlockDevices.unusedBlockDevices"
+		path "org.zstack.header.host.BlockDevices.unusedBlockDevices"
 		desc "没有分区且没有挂载目录或者所有分区没有挂载目录的磁盘"
 		type "List"
 		since "zsv 4.3.0"
@@ -16,7 +16,7 @@ doc {
 	}
 	ref {
 		name "usedBlockDevices"
-		path "org.zstack.header.host.APIGetPhysicalMachineBlockDevicesReply.BlockDevices.usedBlockDevices"
+		path "org.zstack.header.host.BlockDevices.usedBlockDevices"
 		desc "已挂载目录或者有分区挂载目录的磁盘"
 		type "List"
 		since "zsv 4.3.0"
